Changed the get_linked_articles file
Completely rewrote the logic of retrieving all the links of an article. Performance issues should now be resolved, this has been done by using the list of “backlinks” of other articles provided by the Wikipedia API, rather than skimming through each and every linked article in the article we want to search. There is still the need to code the logic for only retrieving the top 30 most relevant articles, but the number of articles should already be limited.